Updated package.json.
Added homepage, keywords, bugs.
Modified description to use the one in haxe.org homepage. Made it clear that it is a wrapper.
Changed license to MIT - this is the license of this package. Licenses of haxe is irrelevant since they're downloaded at install-time.
Added os and cpu, which are depended on the binaries haxe.org provides.